Module Position.Bitboard

A more effecient implementation based on a clever use of bit operations.

type t

Represent a connect4 game position. Columns are 0-based.

val sexp_of_t : t -> Sexplib0.Sexp.t

val width : t -> int

val height : t -> int

val key : [ `not_available | `some of t -> int ]

val create : width:int -> height:int -> t

Create a new position from the given sizes.

val copy : t -> t

Return a deep copy of t, for use in searching algorithm. The two positions can diverge freely from there without affecting each other.

val can_play : t -> column:int -> bool

The given column can be played if it is not currently completely filled.

val play : t -> column:int -> unit

Given a free column, place a token of the next player to play there.

val is_winning_move : t -> column:int -> bool

Indicates whether the current player wins by playing in this column.

val number_of_plies : t -> int

Return the number of plies that have been played from the beginning. This is the number of times play has been called.

val to_ascii_table : t -> string