Module C4solver

module Bench : sig ... end

module Bench_db : sig ... end

A bench db is a local file where we store bench results as we gather them using the bench run command. The motivation for storing them is so that we can show them all together in nice tables, and avoid re-running benches that we have already run too many times unintentionally.

module Bencher : sig ... end

module Gen_scripts : sig ... end

module Measure : sig ... end

module Position : sig ... end

module Solver : sig ... end

module Step : sig ... end

val main : Core.Command.t